requirements.txt
requests>=2.26.0
beautifulsoup4>=4.10.0
python-dotenv>=0.19.0
argparse>=1.4.0
whois>=0.9.5
pytest>=6.2.5
ratelimit>=2.2.1
phonenumbers>=8.12.0
tabulate>=0.8.9
dnspython>=2.1.0